requirements.txt
certifi
-e repos/common-dibbs
pycrypto
python-dateutil
requests
six
urllib3
